requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.2.7
cron-descriptor==1.4.0
defusedxml==0.7.1
Django==4.2.3
django-timezone-field==5.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
flake8==6.0.0
freezegun==1.2.2
gunicorn==20.1.0
httpie==3.2.2
idna==3.4
iniconfig==2.0.0
kombu==5.3.1
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
PySocks==1.7.1
pytest==7.4.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.5.0
XlsxWriter==3.1.2
